Use size_t for array index variables

+++ RHaPSODemo/src/PSO/ParticleSwarm.cpp
@@ -95,7 +95,7 @@ namespace rhapsodies {
 
 		float fPenaltyBest = std::numeric_limits<float>::max();
 		int iIndexBest = 0;
-		for(int particle = 0; particle < m_vecParticles.size(); ++particle) {
+		for(size_t particle = 0; particle < m_vecParticles.size(); ++particle) {
 			if(m_vecParticles[particle].GetIBestPenalty() < fPenaltyBest) {
 				fPenaltyBest = m_vecParticles[particle].GetIBestPenalty();
 				iIndexBest = particle;
@@ -105,7 +105,7 @@ namespace rhapsodies {
 		Particle oParticleBest = m_vecParticles[iIndexBest];
 
 		// static constriction coefficient and behavioral parameters for now
-		for(int particle = 0; particle < m_vecParticles.size(); ++particle) {
+		for(size_t particle = 0; particle < m_vecParticles.size(); ++particle) {
 			m_vecParticles[particle].Imitate(oParticleBest, 2.8f, 1.3f);				
 		}		
 	}
